Specialize scalar SetBufferData for uniform blocks
Assuming the shader has the uniform block configured as std140, the
allocated block must be a multiple of 16. Since we're getting the size
of the allocated block from the data type's size, the data type must
also have a multiple-of-16 size.

The data type's members must also be aligned to a multiple of 16
(otherwise the shader will not read the values correctly); however, I'm
not sure if there's a way to enforce that.

Technically the vector version should also have a specialization, but
I'm still debating how best to handle it. (It might be best to just
disable it for uniform buffers.)